+++ spring-plus-governor/src/main/java/io/github/oatelauser/springplus/governor/idempotent/RepeatSubmitInterceptor.java
@@ -39,7 +39,9 @@ public class RepeatSubmitInterceptor implements MethodInterceptor {
         try {
             return invocation.proceed();
         } catch (Throwable throwable) {
-            store.delete(key);
+            if (annotation.releaseOnFailure()) {
+                store.delete(key);
+            }
             throw throwable;
         }
     }